Move UserModal component to its own file
